+++ CHANGELOG.md
@@ -9,13 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added name and version of program to window title.
 - Added some padding to input elements.
 - Added specific tile type colors to tile buttons.
-- Limited possible level time from 1 second to 120 seconds.
 - Added .json file extension filter and a generated file name to "Export" dialog.
+- Added possibility to change amount of tile rows/columns (default is 8 x 8).
+- Added amount of tile rows/columns to window title.
 
 ### Changed
 - Enclosed and outsourced level export function to "Level" class.
 - Changed window width from 400 pixels to 500 pixels.
 - Changed design of error messages to error messages with icon.
+- Limited possible level time from 1 second to 120 seconds.
 
 ## [1.0.0] - 2018-04-17
 ### Added